update Profile tests
